migretion修正

+++ back/database/migrations/2025_03_13_094315_create_skill_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->string('name')->unique();
             $table->text('description');
             $table->string('effect_type');
-            $table->integer('effect_amount');
+            $table->smallInteger('effect_amount');
             $table->timestamps();
         });
     }
